more webpack changes, got the real build working
