+++ controllers/authentication.js
@@ -1,12 +1,11 @@
 
 // TODO eventually put passport in here too.
 
+const path = require('path')
+
 const auth = {
    login : function(req, res, next) {
-      res.render('base', {
-        app: "login",
-        layout: false
-      });
+      res.sendFile('/dist/index.html', {root : path.resolve('./public') })
    },
    logout : function(req, res, next) {
      req.logout();
