+++ src/index.html
@@ -2,10 +2,16 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <title>vue-test</title>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <link href="https://unpkg.com/vuetify/dist/vuetify.min.css" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css?family=Orbitron" rel="stylesheet">
+    <title>NodeBoard</title>
   </head>
   <body>
-    <div id="app"></div>
+   <div id="app">
+   </div>
     <!-- built files will be auto injected -->
   </body>
 </html>
